Data filtering in MainView improved, random data generation improved

+++ src/main/java/com/example/banksystem/credit/CreditType.java
@@ -19,9 +19,9 @@ public class CreditType {
     @ManyToOne
     private Currency currency;
     @Column(name = "min_amount", length = 8, nullable = false)
-    private int minAmount;
+    private Long minAmount;
     @Column(name = "max_amount", length = 8, nullable = false)
-    private int maxAmount;
+    private Long maxAmount;
     @Column(name = "repayment_time", length = 8, nullable = false)
     private int repaymentTime;
     //
@@ -29,7 +29,7 @@ public class CreditType {
 
     }
 
-    public CreditType(String name, double rate, int minAmount, int maxAmount, int repaymentTime, Currency currency) {
+    public CreditType(String name, double rate, Long minAmount, Long maxAmount, int repaymentTime, Currency currency) {
         this.name = name;
         this.rate = rate;
         this.minAmount = minAmount;
@@ -62,19 +62,19 @@ public class CreditType {
         this.rate = rate;
     }
 
-    public int getMinAmount() {
+    public Long getMinAmount() {
         return minAmount;
     }
 
-    public void setMinAmount(int minAmount) {
+    public void setMinAmount(Long minAmount) {
         this.minAmount = minAmount;
     }
 
-    public int getMaxAmount() {
+    public Long getMaxAmount() {
         return maxAmount;
     }
 
-    public void setMaxAmount(int maxAmount) {
+    public void setMaxAmount(Long maxAmount) {
         this.maxAmount = maxAmount;
     }
